@@ -12,9 +12,9 @@
     },
     "devDependencies": {
         "@sveltejs/adapter-auto": "^3.0.0",
-        "@sveltejs/adapter-node": "^5.2.0",
-        "@sveltejs/kit": "^2.8.3",
-        "@sveltejs/vite-plugin-svelte": "^3.0.0",
+        "@sveltejs/adapter-node": "^5.2.9",
+        "@sveltejs/kit": "^2.0.0",
+        "@sveltejs/vite-plugin-svelte": "^4.0.0",
         "@types/eslint": "^9.6.1",
         "@types/node": "^22.7.4",
         "@types/sanitize-html": "^2.13.0",
@@ -24,15 +24,15 @@
         "eslint-plugin-svelte": "^2.46.1",
         "globals": "^15.12.0",
         "prettier": "^3.1.1",
-        "prettier-plugin-svelte": "^3.1.2",
+        "prettier-plugin-svelte": "^3.2.6",
         "sass": "^1.81.1",
-        "svelte": "^4.2.19",
+        "svelte": "^5.0.0",
         "svelte-check": "^4.1.0",
         "svelte-preprocess": "^6.0.1",
         "tslib": "^2.7.0",
         "typescript": "^5.5.4",
         "typescript-eslint": "^8.19.1",
-        "vite": "^5.4.6"
+        "vite": "^5.4.11"
     },
     "type": "module",
     "dependencies": {
@@ -55,7 +55,7 @@
         "sanitize-html": "^2.14.0",
         "shiki": "^1.14.1",
         "sortablejs": "^1.15.0",
-        "svelte-french-toast": "^1.2.0",
+        "svelte-5-french-toast": "^2.0.1",
         "tippy.js": "^6.3.7",
         "tm-grammars": "^1.22.1"
     }
