@@ -1,20 +1,28 @@
 <script lang="ts">
-    import { onMount, tick } from "svelte";
+    import { onMount, tick, mount, unmount } from "svelte";
     import Sortable, { type SortableEvent } from "sortablejs";
     import Tab from "./Tab.svelte";
     import Editor from "./Editor.svelte";
-    import TabData from "./TabData";
+    import TabData from "./TabData.svelte";
     import { getLangs, type Language } from "./api/lang";
     import { beforeNavigate } from "$app/navigation";
     import { creatingPasteStore } from "./stores";
     import type { Settings } from "./api/settings";
     import type { Pasty } from "./api/paste";
 
-    export let settings: Settings;
-    export let tabs: TabData[] = new Array<TabData>();
-    export let activeTab: TabData | undefined = undefined;
-    export let existingPasties: Pasty[] = [];
-    $: activeTab = tabs.find((t) => t.id === activeTabId);
+    interface Props {
+        settings: Settings;
+        tabs: TabData[];
+        activeTab?: TabData;
+        existingPasties?: Pasty[];
+    }
+
+    let {
+        settings,
+        tabs = $bindable([]),
+        activeTab = $bindable(),
+        existingPasties = []
+    }: Props = $props();
 
     // used for giving tabs their own unique ID
     let tabCounter = 0;
@@ -22,13 +30,17 @@
     let tabGroupElement: HTMLElement;
     let editorTarget: HTMLElement;
 
-    let activeTabId = "0";
+    let activeTabId = $state("0");
 
-    let isDraggedOver = false;
+    let isDraggedOver = $state(false);
+
+    $effect(() => {
+        activeTab = tabs.find((t) => t.id === activeTabId)!;
+    });
 
     beforeNavigate((navigation) => {
-        if ($creatingPasteStore) {
-            $creatingPasteStore = false;
+        if (creatingPasteStore) {
+            creatingPasteStore.set(false);
             return;
         }
 
@@ -77,14 +89,14 @@
         const idx = tabs.findIndex((t) => t.id === id);
 
         // ask if it's okay to close a non empty tab
-        if (tabs[idx].editor.getContent().length > 0) {
+        if (tabs[idx].editor!.getContent().length > 0) {
             if (!confirm("are you sure you want to close a non-empty tab?")) {
                 return;
             }
         }
 
         // destroy editor element
-        tabs[idx].editor.$destroy();
+        unmount(tabs[idx].editor!);
 
         // remove from array
         tabs = [...tabs.slice(0, idx), ...tabs.slice(idx + 1, tabs.length)];
@@ -100,15 +112,17 @@
         updateTabEditorVisibility();
     };
 
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const onTabClick = async (id: string, event: CustomEvent<any>) => {
-        let target = event.detail.event.target as HTMLElement;
+    const onTabClick = async (
+        id: string,
+        event: Event & { currentTarget: EventTarget & Element }
+    ) => {
+        let target = event.target as HTMLElement;
 
         // ignore if close icon is pressed
         if (target.nodeName === "SVG") return;
 
         // ignore if double click, so the rename field doesn't unfocus
-        if (event.detail.event.detail > 1) return;
+        if ((event as unknown as MouseEvent).detail > 1) return;
 
         await setActiveTab(id);
     };
@@ -116,7 +130,7 @@
     const onTabFinishRenaming = (id: string) => {
         const idx = tabs.findIndex((t) => t.id === id);
 
-        tabs[idx].editor.focus();
+        tabs[idx].editor!.focus();
     };
 
     export const addTab = async (
@@ -130,26 +144,26 @@
         let newtab: TabData = new TabData(
             id || String(tabCounter),
             name,
-            new Editor({
+            mount(Editor, {
                 target: editorTarget,
                 props: {
                     settings,
                     onMounted: () => {
                         if (content && language) {
-                            newtab.editor.setContent(content);
-                            newtab.editor.setSelectedLang(language);
+                            newtab.editor?.setContent(content);
+                            newtab.editor?.setSelectedLang(language);
 
                             // if the first tab is empty, remove it
                             const firstTab = tabs[0];
                             if (
                                 tabs.length > 1 &&
-                                !firstTab.editor.getContent() &&
+                                !firstTab.editor?.getContent() &&
                                 firstTab.title === "untitled"
                             ) {
                                 tabs = tabs.slice(1);
                             }
                         } else {
-                            copyPreviousTabSettings(newtab.editor);
+                            copyPreviousTabSettings(newtab.editor!);
                         }
                     }
                 }
@@ -163,12 +177,12 @@
     };
 
     // set the language and indentation same as the previous tab
-    const copyPreviousTabSettings = (newTabEditor: Editor) => {
+    const copyPreviousTabSettings = (newTabEditor: ReturnType<typeof Editor>) => {
         if (tabs.length > 1) {
             const previousEditor = tabs[tabs.length - 2].editor;
 
-            newTabEditor.setSelectedLang(previousEditor.getSelectedLang());
-            newTabEditor.setIndentaion(...previousEditor.getIndentation());
+            newTabEditor.setSelectedLang(previousEditor!.getSelectedLang()!);
+            newTabEditor.setIndentaion(...previousEditor!.getIndentation());
         }
     };
 
@@ -184,13 +198,13 @@
         let tab = tabs.find((t) => t.id === activeTabId);
 
         if (tab && !tab.isInRenamingState) {
-            tab.editor.focus();
+            tab.editor!.focus();
         }
     };
 
     const updateTabEditorVisibility = () => {
         for (let tab of tabs) {
-            tab.editor.$set({ hidden: !(tab.id === activeTabId) });
+            tab.editor?.setHidden(!(tab.id === activeTabId));
         }
     };
 
@@ -228,7 +242,7 @@
 
     const hasModifiedTabs = () => {
         for (const tab of tabs) {
-            if (tab.editor.getContent()) return true;
+            if (tab.editor?.getContent()) return true;
         }
 
         return false;
@@ -236,7 +250,7 @@
 </script>
 
 <svelte:window
-    on:beforeunload={(e) => {
+    onbeforeunload={(e) => {
         if (hasModifiedTabs()) {
             // prevent default only after the check, firefox issue
             e.preventDefault();
@@ -248,14 +262,10 @@
     }}
 />
 
-<svelte:body
-    on:dragover={handleDragOver}
-    on:drop={handleDragDrop}
-    on:dragleave={closeDragContainer}
-/>
+<svelte:body ondragover={handleDragOver} ondrop={handleDragDrop} ondragleave={closeDragContainer} />
 
 <div class="drop-container" class:drop-container--shown={isDraggedOver}>
-    <div class="drop-container__cover" />
+    <div class="drop-container__cover"></div>
     <p>drop files here</p>
 </div>
 
@@ -265,18 +275,19 @@
         {#each tabs as tab, _ (tab.id)}
             <Tab
                 id={tab.id.toString()}
-                on:close={() => onTabClose(tab.id)}
-                on:click={(event) => onTabClick(tab.id, event)}
-                on:finishedRenaming={() => onTabFinishRenaming(tab.id)}
-                bind:isInRenamingState={tab.isInRenamingState}
+                onclose={() => onTabClose(tab.id)}
+                onclick={(event) => onTabClick(tab.id, event)}
+                onfinishedRenaming={() => onTabFinishRenaming(tab.id)}
                 bind:title={tab.title}
+                bind:isInRenamingState={tab.isInRenamingState}
                 isActive={activeTabId === tab.id}
                 closeable={tabs.length > 1}
+                isReadonly={false}
             />
         {/each}
     </div>
 
-    <button class="add-btn btn" on:click={() => addTab()}>
+    <button class="add-btn btn" onclick={() => addTab()}>
         <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" class="icon">
             <title>Plus Icon</title>
             <path
@@ -288,7 +299,7 @@
     </button>
 </div>
 
-<div class="editor" bind:this={editorTarget} />
+<div class="editor" bind:this={editorTarget}></div>
 
 <style lang="scss">
     .drop-container {
