@@ -4,7 +4,11 @@
     import { env } from "$env/dynamic/public";
     import PasteList from "$lib/PasteList.svelte";
 
-    export let data: PageData;
+    interface Props {
+        data: PageData;
+    }
+
+    let { data }: Props = $props();
 </script>
 
 <svelte:head>
